Split Torch classifier example into multiple files
