+++ samples/torch/src/main/kotlin/Disposable.kt
@@ -1,0 +1,26 @@
+interface Disposable {
+    fun dispose()
+}
+
+open class DisposableContainer(private val disposables: MutableList<Disposable> = ArrayList()) : Disposable {
+    /**
+     * Creates the object and schedules its disposal for the end of the scope.
+     */
+    fun <T : Disposable> use(create: () -> T) = create().also { disposables.add(it) }
+
+    override fun dispose() {
+        for (disposable in disposables) {
+            disposable.dispose()
+        }
+    }
+}
+
+fun <T> disposeScoped(action: DisposableContainer.() -> T): T {
+    val scope = DisposableContainer()
+
+    try {
+        return scope.action()
+    } finally {
+        scope.dispose()
+    }
+}
